+++ build.gradle
@@ -38,7 +38,8 @@ dependencies {
 
     modImplementation("net.fabricmc.fabric-api:fabric-api:${project.fabric_version}")
     include(modImplementation(fabricApi.module("fabric-data-generation-api-v1", project.fabric_version))) // include datagen api
-    modImplementation(include("io.github.fabricators_of_create.Porting-Lib:tags:2.0.588+1.19.2"))
+    modImplementation(include("io.github.fabricators_of_create.Porting-Lib:tags:${project.porting_lib_version}"))
+    modImplementation(include("io.github.fabricators_of_create.Porting-Lib:model_generators:${project.porting_lib_version}"))
     implementation(include("com.github.LlamaLad7:MixinExtras:0.0.11"))
     annotationProcessor("com.github.LlamaLad7:MixinExtras:0.0.11")
