+++ gradle.properties
@@ -4,6 +4,7 @@ mod_version = 1.1.6
 minecraft_version = 1.19.2
 loader_version = 0.14.9
 fabric_version = 0.64.0+1.19.2
+porting_lib_version = 2.0.588+1.19.2
 
 # Mappings
 # https://lambdaurora.dev/tools/import_quilt.html
