+++ src/main/java/com/tterrag/registrate/builders/EntityBuilder.java
@@ -154,9 +154,7 @@ public class EntityBuilder<T extends Entity, P> extends AbstractBuilder<EntityTy
             throw new IllegalStateException("Cannot configure attributes more than once");
         }
         attributesConfigured = true;
-        this.onRegister(t -> {
-            FabricDefaultAttributeRegistry.register((EntityType<? extends LivingEntity>) getEntry(), attributes.get());
-        });
+        this.onRegister(entry -> FabricDefaultAttributeRegistry.register((EntityType<? extends LivingEntity>) entry, attributes.get()));
         return this;
     }
