--- src/main/java/net/minecraftforge/client/model/generators/BlockModelBuilder.java
@@ -1,38 +1,0 @@
-/*
- * Minecraft Forge
- * Copyright (c) 2016-2021.
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation version 2.1
- * of the License.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-
-package net.minecraftforge.client.model.generators;
-
-import net.minecraft.resources.ResourceLocation;
-import net.minecraftforge.common.data.ExistingFileHelper;
-
-/**
- * Builder for block models, does not currently provide any additional
- * functionality over {@link ModelBuilder}, purely a stub class with a concrete
- * generic.
- * 
- * @see ModelProvider
- * @see ModelBuilder
- */
-public class BlockModelBuilder extends ModelBuilder<BlockModelBuilder> {
-
-    public BlockModelBuilder(ResourceLocation outputLocation, ExistingFileHelper existingFileHelper) {
-        super(outputLocation, existingFileHelper);
-    }
-}
