--- src/main/java/net/minecraftforge/client/model/generators/BlockModelProvider.java
@@ -1,42 +1,0 @@
-/*
- * Minecraft Forge
- * Copyright (c) 2016-2021.
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation version 2.1
- * of the License.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-
-package net.minecraftforge.client.model.generators;
-
-import org.jetbrains.annotations.NotNull;
-
-import net.minecraft.data.DataGenerator;
-import net.minecraftforge.common.data.ExistingFileHelper;
-
-/**
- * Stub class to extend for block model data providers, eliminates some
- * boilerplate constructor parameters.
- */
-public abstract class BlockModelProvider extends ModelProvider<BlockModelBuilder> {
-
-    public BlockModelProvider(DataGenerator generator, String modid, ExistingFileHelper existingFileHelper) {
-        super(generator, modid, BLOCK_FOLDER, BlockModelBuilder::new, existingFileHelper);
-    }
-
-    @NotNull
-    @Override
-    public String getName() {
-        return "Block Models: " + modid;
-    }
-}
