--- src/main/java/net/minecraftforge/client/model/generators/ItemModelBuilder.java
@@ -1,103 +1,0 @@
-/*
- * Minecraft Forge
- * Copyright (c) 2016-2021.
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation version 2.1
- * of the License.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-
-package net.minecraftforge.client.model.generators;
-
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-
-import com.google.common.base.Preconditions;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
-
-import net.minecraft.resources.ResourceLocation;
-import net.minecraftforge.common.data.ExistingFileHelper;
-
-/**
- * Builder for item models, adds the ability to build overrides via
- * {@link #override()}.
- */
-public class ItemModelBuilder extends ModelBuilder<ItemModelBuilder> {
-
-    protected List<OverrideBuilder> overrides = new ArrayList<>();
-
-    public ItemModelBuilder(ResourceLocation outputLocation, ExistingFileHelper existingFileHelper) {
-        super(outputLocation, existingFileHelper);
-    }
-
-    public OverrideBuilder override() {
-        OverrideBuilder ret = new OverrideBuilder();
-        overrides.add(ret);
-        return ret;
-    }
-
-    /**
-     * Get an existing override builder
-     * 
-     * @param index the index of the existing override builder
-     * @return the override builder
-     * @throws IndexOutOfBoundsException if {@code} index is out of bounds
-     */
-    public OverrideBuilder override(int index) {
-        Preconditions.checkElementIndex(index, overrides.size(), "override");
-        return overrides.get(index);
-    }
-
-    @Override
-    public JsonObject toJson() {
-        JsonObject root = super.toJson();
-        if (!overrides.isEmpty()) {
-            JsonArray overridesJson = new JsonArray();
-            overrides.stream().map(OverrideBuilder::toJson).forEach(overridesJson::add);
-            root.add("overrides", overridesJson);
-        }
-        return root;
-    }
-
-    public class OverrideBuilder {
-
-        private ModelFile model;
-        private final Map<ResourceLocation, Float> predicates = new LinkedHashMap<>();
-
-        public OverrideBuilder model(ModelFile model) {
-            this.model = model;
-            model.assertExistence();
-            return this;
-        }
-
-        public OverrideBuilder predicate(ResourceLocation key, float value) {
-            this.predicates.put(key, value);
-            return this;
-        }
-
-        public ItemModelBuilder end() { return ItemModelBuilder.this; }
-
-        JsonObject toJson() {
-            JsonObject ret = new JsonObject();
-            JsonObject predicatesJson = new JsonObject();
-            predicates.forEach((key, val) -> predicatesJson.addProperty(key.toString(), val));
-            ret.add("predicate", predicatesJson);
-            ret.addProperty("model", model.getLocation().toString());
-            return ret;
-        }
-    }
-
-}
